+++ src/AppBundle/Controller/API/ProjectScopeController.php
@@ -4,6 +4,7 @@ namespace AppBundle\Controller\API;
 
 use AppBundle\Entity\ProjectScope;
 use AppBundle\Form\ProjectScope\CreateType;
+use AppBundle\Security\ProjectVoter;
 use MainBundle\Controller\API\ApiController;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
@@ -78,6 +79,8 @@ class ProjectScopeController extends ApiController
      */
     public function getAction(ProjectScope $projectScope)
     {
+        $this->denyAccessUnlessGranted(ProjectVoter::VIEW, $projectScope->getProject());
+
         return $this->createApiResponse($projectScope);
     }
 
@@ -94,6 +97,8 @@ class ProjectScopeController extends ApiController
      */
     public function editAction(Request $request, ProjectScope $projectScope)
     {
+        $this->denyAccessUnlessGranted(ProjectVoter::EDIT, $projectScope->getProject());
+
         $data = $request->request->all();
         $form = $this->createForm(CreateType::class, $projectScope, ['csrf_protection' => false]);
         $form->submit($data, false);
@@ -128,6 +133,8 @@ class ProjectScopeController extends ApiController
      */
     public function deleteAction(ProjectScope $projectScope)
     {
+        $this->denyAccessUnlessGranted(ProjectVoter::DELETE, $projectScope->getProject());
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($projectScope);
         $em->flush();
